reduce redundant code
