+++ fastdeploy/engine/common_engine.py
@@ -18,7 +18,6 @@ from __future__ import annotations
 
 import asyncio
 import copy
-import glob
 import json
 import multiprocessing
 import os
@@ -68,7 +67,7 @@ from fastdeploy.splitwise.internal_adapter_utils import InternalAdapter
 from fastdeploy.splitwise.splitwise_connector import SplitwiseConnector
 from fastdeploy.trace.constants import LoggingEventName
 from fastdeploy.trace.trace_logger import print as trace_print
-from fastdeploy.utils import EngineError, console_logger, envs, get_logger, llm_logger
+from fastdeploy.utils import EngineError, console_logger, ensure_workerlog_alias, envs, get_logger, llm_logger
 
 try:
     TokenProcessor = load_token_processor_plugins()
@@ -1908,36 +1907,6 @@ class EngineService:
             command_prefix += f"{k}={v} "
         return command_prefix
 
-    def _ensure_workerlog_alias(self, log_dir: str, paddle_log_dir: str) -> None:
-        try:
-            def ensure_alias_for_target(target_path: str) -> None:
-                if not target_path:
-                    return
-                link_path = os.path.join(log_dir, os.path.basename(target_path))
-                abs_target_path = os.path.abspath(target_path)
-                if os.path.islink(link_path):
-                    current_target = os.readlink(link_path)
-                    abs_current_target = os.path.abspath(os.path.join(os.path.dirname(link_path), current_target))
-                    if abs_current_target == abs_target_path:
-                        return
-                    os.remove(link_path)
-                elif os.path.isfile(link_path):
-                    os.remove(link_path)
-                elif os.path.exists(link_path):
-                    return
-                os.symlink(abs_target_path, link_path)
-
-            # Always alias workerlog.0 even if it does not exist yet.
-            ensure_alias_for_target(os.path.join(paddle_log_dir, "workerlog.0"))
-
-            # Alias any existing workerlog.* files.
-            for target_path in glob.glob(os.path.join(paddle_log_dir, "workerlog.*")):
-                if os.path.isfile(target_path):
-                    ensure_alias_for_target(target_path)
-        except OSError:
-            # Best effort: if symlink creation fails, keep paddle logs intact.
-            pass
-
     def _start_worker_service(self):
         """
         start gpu worker service
@@ -1946,7 +1915,7 @@ class EngineService:
         log_dir = os.getenv("FD_LOG_DIR", default="log")
         paddle_log_dir = os.path.join(log_dir, "paddle")
         os.makedirs(paddle_log_dir, exist_ok=True)
-        self._ensure_workerlog_alias(log_dir, paddle_log_dir)
+        ensure_workerlog_alias(log_dir, paddle_log_dir)
         command_prefix = self._setting_environ_variables()
         current_file_path = os.path.abspath(__file__)
         current_dir_path = os.path.split(current_file_path)[0]
